+++ src/CommonDBTM.php
@@ -1280,7 +1280,7 @@ class CommonDBTM extends CommonGLPI
     {
         global $CFG_GLPI, $DB;
 
-        if ($DB->isSlave()) {
+        if ($DB->isReplica()) {
             return false;
         }
 
@@ -1632,7 +1632,7 @@ class CommonDBTM extends CommonGLPI
     {
         global $DB;
 
-        if ($DB->isSlave()) {
+        if ($DB->isReplica()) {
             return false;
         }
 
@@ -2108,7 +2108,7 @@ class CommonDBTM extends CommonGLPI
     {
         global $DB;
 
-        if ($DB->isSlave()) {
+        if ($DB->isReplica()) {
             return false;
         }