+++ src/Glpi/Features/Clonable.php
@@ -255,7 +255,7 @@ trait Clonable
     {
         global $DB;
 
-        if ($DB->isSlave()) {
+        if ($DB->isReplica()) {
             return false;
         }
         $new_item = new static();